move Timezone update UI to Preference
